### 2022-11-11 #4925 @editorialbot
[PRE REVIEW]: py-sc-fermi: self-consistent Fermi energies and defect concentrations from electronic structure calculations
**Submitting author:** @alexsquires (<a href="http://orcid.org/0000-0001-6967-3690">Alexander G. Squires</a>)
**Repository:** https://github.com/bjmorgan/py-sc-fermi
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.1
