### 2022-11-17 #4925 @rkurchin
👋 @agoyal06, hope your new job is treating you well! Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
