### 2022-11-19 #4925 @agoyal06
Thanks for sharing, Rachel. Hope you are doing well.

I will be interested in reviewing. I am Currently on travel with limited connectivity, so it may take longer to review. By when you need the review?

Get Outlook for Android<https://aka.ms/AAb9ysg>
