### 2022-11-19 #4925 @hongzhouye
Hi Rachel,

Thank you for the invitation! Yes, I am interested in reviewing it if it's
still available.
Since I'm new to this, is there a guideline for reviewers? (I found a
checklist <https://joss.readthedocs.io/en/latest/review_checklist.html> here
which looks very useful; just in case there are more specific ones). And
roughly
by when do you want my comments? Thanks!

Best,
Hong-Zhou

On Thu, Nov 17, 2022 at 9:14 AM Rachel Kurchin ***@***.***>
wrote:

> No problem, @xwang862 <https://github.com/xwang862>, thanks for letting
> me know. @hongzhouye <https://github.com/hongzhouye>, if you're
> available, we'd be happy to have you as a reviewer!
>
