+++ paper/paper.bib
@@ -7,6 +7,7 @@ volume  ="13",
 issue  ="5",
 pages  ="1481-1491",
 publisher  ="The Royal Society of Chemistry",
+doi = "10.1039/d0ee00291g"
 }
 
 @ARTICLE{Maier2003-pc,
@@ -20,7 +21,8 @@ publisher  ="The Royal Society of Chemistry",
   pages     = "2164--2173",
   month     =  may,
   year      =  2003,
-  language  = "en"
+  language  = "en",
+  doi = "10.1039/b300139n"
 }
 
 @ARTICLE{Toriyama2021-ch,
@@ -34,7 +36,8 @@ publisher  ="The Royal Society of Chemistry",
   pages     = "20685--20694",
   month     =  sep,
   year      =  2021,
-  language  = "en"
+  language  = "en",
+  doi: "10.1039/D1TA05112A"
 }
 
 @ARTICLE{Shimoda2022-vs,
@@ -62,7 +65,8 @@ publisher  ="The Royal Society of Chemistry",
   number    =  2,
   pages     = "496--500",
   month     =  feb,
-  year      =  2018
+  year      =  2018,
+  doi = "10.1021/acsenergylett.7b01313.s001"
 }
 
 @article{batteries,
@@ -89,7 +93,8 @@ publisher  ="The Royal Society of Chemistry",
   pages    = "7217--7227",
   month    =  may,
   year     =  2021,
-  language = "en"
+  language = "en",
+  doi = "10.1021/acs.inorgchem.1c00278"
 }
 
 @Article{TCOs,
@@ -101,7 +106,8 @@ publisher  ="The Royal Society of Chemistry",
   number    =  36,
   pages     = "11995--12009",
   year      =  2021,
-  language  = "en"
+  language  = "en",
+  doi = "10.1039/d1tc02547c"
 }
 
 @ARTICLE{Neilson2022-cj,
@@ -113,7 +119,8 @@ publisher  ="The Royal Society of Chemistry",
   pages    = "111434",
   month    =  jul,
   year     =  2022,
-  keywords = "Defect chemistry; Thermodynamics; Brouwer diagrams"
+  keywords = "Defect chemistry; Thermodynamics; Brouwer diagrams",
+  doi = "10.1016/j.commatsci.2022.111434"
 }
 
 @ARTICLE{Buckeridge2019-fm,
@@ -123,7 +130,8 @@ publisher  ="The Royal Society of Chemistry",
   author    = "Buckeridge, J",
   journal   = "Comput. Phys. Commun.",
   publisher = "Elsevier",
-  year      =  2019
+  year      =  2019,
+  doi = "10.1016/j.cpc.2019.06.017"
 }
 
 @ARTICLE{Ogawa2022-sn,
@@ -137,7 +145,8 @@ publisher  ="The Royal Society of Chemistry",
   pages     = "1--11",
   month     =  apr,
   year      =  2022,
-  language  = "en"
+  language  = "en",
+  doi = "10.1038/s41524-022-00756-0"
 }
 
 @ARTICLE{Arrigoni2021-oc,
@@ -150,7 +159,8 @@ publisher  ="The Royal Society of Chemistry",
   month    =  jul,
   year     =  2021,
   keywords = "First-principles; Point defects; Charged defects; Thermodynamic
-              stability; Density functional theory; Python"
+              stability; Density functional theory; Python",
+  doi = "10.1016/j.cpc.2021.107946"
 }
 
 @ARTICLE{LLZO-elect,
@@ -165,7 +175,8 @@ publisher  ="The Royal Society of Chemistry",
   month     =  aug,
   year      =  2022,
[…]
