+++ paper/paper.bib
@@ -37,7 +37,7 @@ doi = "10.1039/d0ee00291g"
   month     =  sep,
   year      =  2021,
   language  = "en",
-  doi: "10.1039/D1TA05112A"
+  doi = "10.1039/D1TA05112A"
 }
 
 @ARTICLE{Shimoda2022-vs,
