+++ paper/paper.bib
@@ -52,7 +52,8 @@ doi = "10.1039/d0ee00291g"
   number    =  12,
   pages     = "5634--5643",
   month     =  jun,
-  year      =  2022
+  year      =  2022,
+  doi = "10.1021/acs.chemmater.2c00944"
 }
 
 @ARTICLE{Kim2018-eg,
@@ -78,6 +79,7 @@ doi = "10.1039/d0ee00291g"
 	author = {Joachim Maier},
 	title = {Review{\textemdash} Battery Materials: Why Defect Chemistry?},
 	journal = {Journal of The Electrochemical Society},
+	doi = "10.1149/2.0011514jes"
 }
 
 @ARTICLE{Haouari2021-xz,
@@ -283,5 +285,6 @@ doi = "10.1039/d0ee00291g"
   author    = "Squires, A G and Scanlon, D O and Morgan, B J",
   journal   = "Chem. Mater.",
   publisher = "ACS Publications",
-  year      =  2019
+  year      =  2019,
+  doi = "10.1021/acs.chemmater.9b04319"
 }
