### 2022-11-28 c2a2ebd
fix spin pol when parsing dos from two arrays
