### 2022-11-30 71a55e9
vasprun parsing consistent with dict and list of lists for DOS
