### 2022-11-30 b2012aa
fix type checkings
