### 2022-12-09 #4962 @hongzhouye
@rkurchin Hi Rachel, I have a very quick question about the "Contribution and authorship". Does every author need to have a recorded contribution to the code (i.e., through the commit history)? Thanks!
