### 2022-12-09 #4962 @rkurchin
@hongzhouye thanks for the thoughtful question! I'm actually not certain if we have a formal policy about this. I've asked the rest of the editorial team and will let you know what I find out ASAP.
