### 2022-12-09 #4962 @hongzhouye
@rkurchin Thank you! In case a bit more details would be helpful, I was particularly asking a case where someone (often a PI) helps conceptualize the work but is not actively involved in the actual coding work.
