### 2022-12-09 #4962 @rkurchin
Okay, the answer is no, there doesn't need to be a commit history contribution – the situation you describe (as well as other conceivable ones) is perfectly fine to merit authorship. Thanks again for checking!
