### 2022-12-09 #4962 @alexsquires
Hi both, when preparing the code for submission, I was checked the authorship policy as I am aware David Scanlon is a "non-commiting" author!

The policy on the website says

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

David's contributions have been scientific, and his knowledge and experience has been invaluable in identifying "edge-cases" where the code was giving numerically stable, but chemically unreasonable self-consistent Fermi energies. This led to a number changes in which the code was structured; it was my feeling from the authorship guidance that to include him would therefore be appropriate.
