### 2022-12-19 bjmorgan/py-sc-fermi#20 @hongzhouye
"The DOS object" section in the API tutorial 
*This issue is part of the review for the Journal of Open Source Software* ([*link*](https://github.com/openjournals/joss-reviews/issues/4962))

Hi developers, when going through the (great) tutorial, I found that [the section on DOS](https://py-sc-fermi.readthedocs.io/en/latest/tutorial.html#the-DOS-object) may be a bit challenging to follow for someone who's not familiar with the Materials Project. Particularly, is it possible to download the said DOS file directly from the linked page [mp-22862](https://materialsproject.org/materials/mp-22862) or one has to go through e.g., their Python API? It seems the latter. And if so, I was wondering if it will be possible to have the downloaded DOS file part of the example (assuming it's not too big).
