### 2022-12-19 bjmorgan/py-sc-fermi#20 @alexsquires
Hi Hongzhou,

The idea of the tutorial is not such that the data should be downloaded from the materials project, but that it is provided in the examples directory under "dos.csv", is this unclear? The Materials project is only mentioned as the source!
