### 2022-12-19 bjmorgan/py-sc-fermi#20 @hongzhouye
Hi Alex,

Thanks for your quick reply! After a `find . -name "*csv"` search now I see that it is located in `docs/source`. I was looking for it in the `tests` directory (which is the closest I could find to an "examples" directory). I will close the issue. Thanks again!
