### 2022-12-20 #4962 @agoyal06
Hi Rachel,

Sincere apologies for the delay on my side. I joined my new institute past week and it was difficult for me to get the review done.

Is anything still needed from me?

Regards,
Anuj

Get Outlook for Android<https://aka.ms/AAb9ysg>
