### 2023-01-12 08cde1d
add as_dict() behavior
