### 2023-01-13 f9a21f7
attempt to fix json encoding for DefectChargeState and DefectSpecies
