### 2023-01-13 04b7898
Update defect_system.py
