### 2023-01-18 6ae9840
Update defect_system.py
