### 2023-01-18 d008ab1
Update usage_notes.rst
