### 2023-01-22 ad27503
Update usage_notes.rst
