### 2023-01-24 #4962 @rkurchin
Minor editorial comments:
* line 13: "obtained" is duplicated
* line 14: you can probably remove "calculations" since you said that earlier in the sentence
* lines 11 and 32-33: be consistent about hyphenation in phrase "net charge neutrality" (I would suggest either no hyphens at all, or if you read it as "the neutrality of net charge" as opposed to "charge neutrality in a net fashion" then perhaps "net-charge neutrality"...if you want a deeper dive than necessary on this distinction, [check this out](https://www.texasbarpractice.com/2021/09/07/grammar-tip-hyphenating-compound-adjectives/)
* 37: I assume it's supposed to be "iterative" (or maybe "interactive"?) but in any case "interative" is not a word 😉
* line 66: I would turn the comma into a period and break the "in addition" clause out to a new sentence

Broader conceptual comment: I think it's worth clarifying at least once that there's an underlying presumption of thermodynamic equilibrium in (most of) this. My suggestion would be simply modifying the end of the first sentence of the summary to read "...under the constraint of net charge neutrality and assumption of thermodynamic (quasi-)equilibrium" and perhaps also add "equilibrium" before "concentrations" on line 30. You could then perhaps add one more clarifying allusion to this idea in the paragraph beginning on line 68, since that's a nice bit of functionality for incorporating non-equilibrium information, if/when it is known.

Overall, this looks like a great little tool, and I'll definitely be keeping it in mind for my own work!
