+++ paper/paper.md
@@ -32,25 +32,26 @@ bibliography: paper.bib
 
 # Summary
 
-`py-sc-fermi` is a Python package for calculating point defect concentrations in crystalline materials, under the constraint of net&ndash;charge-neutrality.
+`py-sc-fermi` is a Python package for calculating point defect concentrations in crystalline materials, under the constraint of net&ndash;charge-neutrality and the assumption of thermodynamic (quasi-)equilibrium.
 The required inputs are the formation energies of all point defects of interest and the electronic density of states.
-These can be obtained obtained from electronic structure calculations, e.g., density functional theory calculations.
+These can be obtained from electronic structure calculations, e.g., density functional theory.
 
-Point defects are atomic-scale imperfections crystalline materials, which can strongly affect a range of material properties, including electronic structure [@Pastor2022-xa], charge transport [@batteries], and energy-conversion processes [@TLC].
+Point defects are atomic-scale imperfections in crystalline materials, which can strongly affect a range of material properties, including electronic structure [@Pastor2022-xa], charge transport [@batteries], and energy-conversion processes [@TLC].
 Point defect concentrations are generally sensitive to synthesis conditions and to the inclusion of extrinsic dopants.
 By understanding the thermodynamics of point-defect formation, it is possible to predict the response of specific material properties to changes in synthesis or doping protocols [@thermoelectric; @Squires2019-is; @Park2018-np].
-Attempts to quantitatively model point-defect populations in functional materials has become common practice in the materials modelling community, often with the aim of informing the design of novel, highly performant materials [@Jackson2022-la; @Ganose2018-eu; @Toriyama2021-ch] and to rationlise and optimise the properties of known materials [@Kim2018-eg; @Shimoda2022-vs; @LLZO-elect]. 
+The quantitative modelling of point-defect populations in functional materials has become common practice in the materials modelling community, often undertaken with the aim of informing the design of novel, highly performant materials [@Jackson2022-la; @Ganose2018-eu; @Toriyama2021-ch] or to rationlise and optimise the properties of known materials [@Kim2018-eg; @Shimoda2022-vs; @LLZO-elect]. 
 
 While the formation energies of charge-neutral point defects can be calculated simply, e.g., from electronic structure calculations, the formation energies of charged defects depend on the electronic chemical potential (Fermi energy), which, in turn, depends on the net population of all charged defects in the system.
-The concentrations of charged point defects are therefore mutually dependent, and must be solved as a unified self-consistent system.
-In practice, self-consistent defect concentrations are calculated by imposing the constraint of net-charge neutrality [@Buckeridge2019-fm],
+The equilibrium concentrations of charged point defects are therefore mutually dependent, and must be solved as a unified self-consistent system.
+In practice, self-consistent defect concentrations are calculated by imposing the constraint of net&ndash;charge-neutrality [@Buckeridge2019-fm],
 \begin{equation}
 0 = \sum_{X,q} q \left[ X,q \right] - n_0 + p_0.
 \end{equation}
-Here, the first term is the sum over charge contributions from each defect $X$ in each of its accessible charge states $q$, and $n_0$ and $p_0$ are the concentrations of free electrons and electron holes respectively.
+Here, the first term is a sum over charge contributions from each defect $X$ in each of its accessible charge states $q$.
+$n_0$ and $p_0$ are the concentrations of free electrons and electron holes respectively.
 
-`py-sc-fermi` provides a Python implementation of an interative numerical approach for calculating the self-consistent solution for an arbitrary set of point-defects.
[…]
