### 2023-01-25 #4962 @ajgoyal
Hi @rkurchin, thank you for the opportunity to review. I found the concept of JOSS very useful. I apologize for the delay from my side.

Congratulations to the authors!
