+++ CITATION.cff
@@ -8,7 +8,7 @@ authors:
   given-names: "Benjamin J."
   orcid: "https://orcid.org/0000-0002-3056-8233"
 title: "py-sc-Fermi"
-version: 1.0.0
+version: 1.0.2
 doi: 10.5281/zenodo.7566489
 repository-code: https://github.com/bjmorgan/py-sc-fermi
 license: MIT
