+++ CITATION.cff
@@ -10,20 +10,20 @@ authors:
 title: "py-sc-Fermi"
 version: 1.0.0
 doi: 10.5281/zenodo.7566489
-repository-code: https://github.com/bjmorgan/py-sc-fermi
+url: https://github.com/bjmorgan/py-sc-fermi
 license: MIT
-references:
-  - type: article:
-    authors:
-      - family-names: "Squires"
-        given-names: "Alexander G."
-      - family-names: "Scanlon"
-        given-names: "David O."
-      - family-names: "Morgan"
-        given-names: "Benjamin J."
-    title: "py-sc-fermi: self-consistent Fermi energies and defect concentrations from electronic structure calculations"
-    year: 2023
-    journal: Journal of Open Source Software
-    doi: 10.21105/joss.04962
+preferred-citation:
+  type: article:
+  authors:
+    - family-names: "Squires"
+      given-names: "Alexander G."
+    - family-names: "Scanlon"
+      given-names: "David O."
+    - family-names: "Morgan"
+      given-names: "Benjamin J."
+  title: "py-sc-fermi: self-consistent Fermi energies and defect concentrations from electronic structure calculations"
+  year: 2023
+  journal: "Journal of Open Source Software"
+  doi: "10.21105/joss.04962"
    
    
