### 2023-01-25 #4962 @bjmorgan
> Great, thanks everyone! Authors, I'll do an editorial pass over the manuscript and send any comments shortly. In the meantime, the next steps for you are:
>
> 1. Merge any and all changes from this review into your main branch and **issue a new version tag**. _(If you want to merge in the paper, you may, but it is not required that the actual manuscript live into the repo in perpetuity since JOSS will host it and you can simply add a badge link or whatever you like. But the actual changes to software and docs do need to be merged!)_
> 2. Create a DOI for the contents of the repo at the same commit corresponding to that version tag, e.g. using figshare or Zenodo. Please make sure that the metadata (version number, title, author list, etc.) match those of your manuscript.
> 3. Post a comment here with the version number and DOI.

- The new version is tag number `1.0.0` → https://github.com/bjmorgan/py-sc-fermi/releases/tag/1.0.0
- Version `1.0.0` is archived at Zenodo here: → 10.5281/zenodo.7567782
