### 2023-01-25 #4962 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
