### 2023-01-26 #4962 @bjmorgan
We have made minor edits to the paper in https://github.com/bjmorgan/py-sc-fermi/commit/b2e420715489deb415800bd4f7a92c8718c66679 and https://github.com/bjmorgan/py-sc-fermi/commit/fadabc716b3ae0a4e05f0c95a8d6ce257689d7dd.

Apart from these we are happy with the final proof.
