### 2023-02-02 #4962 @alexsquires
addressed in [f57479f](https://github.com/bjmorgan/py-sc-fermi/commit/f57479f2d204e5f3543b4668eb507d2120a9e1c3)
